+++ src/components/notion/profile.tsx
@@ -9,13 +9,13 @@ const renderProfileBlock = (block: BlockWithChildren): React.ReactNode => {
         case "heading_1":
             return (
                 <h2 key={block.id} className="text-3xl md:text-5xl mt-4 py-4">
-                    {renderText(block.id, block.heading_1.text)}
+                    {renderText(block.id, block.heading_1.rich_text)}
                 </h2>
             );
         case "paragraph":
             return (
                 <p key={block.id} className="py-4">
-                    {renderText(block.id, block.paragraph.text)}
+                    {renderText(block.id, block.paragraph.rich_text)}
                 </p>
             );
         default: