+++ pages/blog/[id].tsx
@@ -2,12 +2,13 @@ import { Tag } from "@components/card";
 import { Layout } from "@components/layout";
 import { CustomLink } from "@components/link";
 import CustomArticleJsonLd from "@components/meta";
-import { RenderedPageContent } from "@components/notion";
+import { PostContent } from "@components/notion";
 import { Section } from "@components/section";
 
 import { getMonthAndYear } from "@util/datetime";
 import { BlockWithChildren, IPost, ITag } from "@util/interface";
 import { getBlogPosts, getPostBlocks, readPost } from "@util/notion";
+import { getDomainName } from "@util/router";
 
 import { GetStaticPaths, GetStaticProps, NextPage } from "next";
 import { NextSeo } from "next-seo";
@@ -109,7 +110,7 @@ const PostPage: NextPage<PostProps> = ({ post, blocks }: PostProps) => {
                             <div className="flex flex-row items-center text-sm my-2">
                                 <FaLink className="text-jet dark:text-cultured mx-1" />
                                 <CustomLink href={post.link}>
-                                    Read this elsewhere
+                                    Also published at {getDomainName(post.link)}
                                 </CustomLink>
                             </div>
                         )}
@@ -117,7 +118,7 @@ const PostPage: NextPage<PostProps> = ({ post, blocks }: PostProps) => {
                             {post.description}
                         </p>
                     </div>
-                    <RenderedPageContent blocks={blocks} />
+                    <PostContent blocks={blocks} />
                 </div>
             </Section>
         </Layout>