+++ src/components/notion/text.tsx
@@ -72,7 +72,7 @@ export const Paragraph: React.FC<ParagraphBlockProps> = ({
     id,
     paragraph,
 }: ParagraphBlockProps) => {
-    return <p>{renderText(id, paragraph.text)}</p>;
+    return <p>{renderText(id, paragraph.rich_text)}</p>;
 };
 
 type QuoteBlockProps = PropsWithChildren<QuoteBlock>;
@@ -85,7 +85,7 @@ export const Quote: React.FC<QuoteBlockProps> = ({
 }: QuoteBlockProps) => {
     return (
         <blockquote className="ml-3 my-2 font-playfair italic bg-primary text-xl">
-            {has_children ? children : renderText(id, quote.text)}
+            {has_children ? children : renderText(id, quote.rich_text)}
         </blockquote>
     );
 };
@@ -106,7 +106,7 @@ export const Callout: React.FC<CalloutProps> = ({
         <div className="mx-auto flex flex-col md:flex-row items-center rounded-lg font-light px-6 py-2 my-2 md:w-5/6 border-2 border-primary">
             <div>{icon}</div>
             <p className="ml-4">
-                {has_children ? children : renderText(id, callout.text)}
+                {has_children ? children : renderText(id, callout.rich_text)}
             </p>
         </div>
     );