+++ src/util/router.ts
@@ -31,3 +31,22 @@ export const getBaseURL = (): string => {
           }[process.env.NODE_ENV];
     return baseUrl;
 };
+
+export const getDomainName = (url: string): string => {
+    const domainSelectionRe =
+        /\bhttps?:\/\/(?:www\.|ww2\.)?((?:[\w-]+\.){1,}\w+)\b/gm;
+    let domainName = "dev.to"; // Default case
+    let matches;
+    while ((matches = domainSelectionRe.exec(url)) !== null) {
+        // This is necessary to avoid infinite loops with zero-width matches
+        if (matches.index === domainSelectionRe.lastIndex) {
+            domainSelectionRe.lastIndex++;
+        }
+
+        // The result can be accessed through the `m`-variable.
+        matches.forEach((match) => {
+            domainName = match;
+        });
+    }
+    return domainName;
+};